make @hashintel/design-system publishable again, add changeset
