+++ .changeset/grumpy-radios-occur.md
@@ -1,0 +1,5 @@
+---
+"@hashintel/design-system": patch
+---
+
+add data type selector, more icons
