+++ libs/@hashintel/design-system/src/autocomplete-dropdown.tsx
@@ -1,8 +1,8 @@
 import type { PaperProps } from "@mui/material";
 import { autocompleteClasses, Box, Paper } from "@mui/material";
 
-import { popperPlacementSelectors } from "./popper-placement-modifier";
-import { textFieldBorderRadius } from "./theme";
+import { popperPlacementSelectors } from "./popper-placement-modifier.js";
+import { textFieldBorderRadius } from "./theme.js";
 
 export const AutocompleteDropdown = ({
   sx,
