+++ libs/@hashintel/design-system/src/callout.tsx
@@ -7,7 +7,7 @@ import type { SxProps, Theme } from "@mui/material";
 import { Stack, Typography } from "@mui/material";
 import type { PropsWithChildren, ReactNode } from "react";
 
-import { FontAwesomeIcon } from "./fontawesome-icon";
+import { FontAwesomeIcon } from "./fontawesome-icon.js";
 
 type CalloutProps = {
   children: ReactNode;
