+++ libs/@hashintel/design-system/src/components.ts
@@ -1,112 +1,112 @@
-export * from "./alert-modal";
-export * from "./autocomplete";
-export * from "./autocomplete-dropdown";
-export * from "./avatar";
-export * from "./button";
-export * from "./callout";
-export * from "./chip";
-export * from "./chip-group";
-export * from "./data-type-selector";
-export * from "./e-chart";
-export * from "./entity-or-type-icon";
-export * from "./fa-icons/icons";
-export * from "./fontawesome-icon";
-export * from "./form-inline";
-export * from "./icon-angle-right-regular";
-export * from "./icon-arrow-down-left-and-arrow-up-right-to-center";
-export * from "./icon-arrow-down-regular";
-export * from "./icon-arrow-left";
-export * from "./icon-arrow-right-regular";
-export * from "./icon-arrow-rotate-left";
-export * from "./icon-arrow-up-regular";
-export * from "./icon-arrow-up-right";
-export * from "./icon-arrow-up-right-and-arrow-down-left-from-center";
-export * from "./icon-arrow-up-right-from-square-regular";
-export * from "./icon-arrow-up-right-regular";
-export * from "./icon-arrow-up-wide-short-light";
-export * from "./icon-arrows-rotate-regular";
-export * from "./icon-asterisk-regular";
-export * from "./icon-at-regular";
-export * from "./icon-barcode";
-export * from "./icon-bell-light";
-export * from "./icon-bug";
-export * from "./icon-bullseye-light";
-export * from "./icon-button";
-export * from "./icon-caret-down-solid";
-export * from "./icon-check";
-export * from "./icon-check-regular";
-export * from "./icon-chrome";
-export * from "./icon-circle-check-regular";
-export * from "./icon-circle-ellipsis-regular";
-export * from "./icon-circle-nodes-light";
-export * from "./icon-circle-one-regular";
-export * from "./icon-circle-plus";
-export * from "./icon-clock-regular";
-export * from "./icon-close";
-export * from "./icon-code";
-export * from "./icon-copy-regular";
-export * from "./icon-dash";
-export * from "./icon-diagram-nested-light";
-export * from "./icon-diagram-regular";
-export * from "./icon-discord";
-export * from "./icon-download-regular";
-export * from "./icon-eye-regular";
-export * from "./icon-eye-slash-regular";
-export * from "./icon-eye-solid";
-export * from "./icon-feather-regular";
-export * from "./icon-file-regular";
-export * from "./icon-file-spreadsheet-regular";
-export * from "./icon-file-spreadsheet-solid";
-export * from "./icon-forward-step-solid";
-export * from "./icon-graph";
-export * from "./icon-image-regular";
-export * from "./icon-image-solid";
-export * from "./icon-infinity-light";
-export * from "./icon-input-pipe";
-export * from "./icon-input-pipe-regular";
-export * from "./icon-lightbulb-on-rainbow";
-export * from "./icon-lightbulb-on-regular";
-export * from "./icon-link";
-export * from "./icon-list-regular";
-export * from "./icon-magnifying-glass-minus-light";
-export * from "./icon-magnifying-glass-plus-light";
-export * from "./icon-magnifying-glass-regular";
-export * from "./icon-memo-circle-check-regular";
-export * from "./icon-microscope-regular";
-export * from "./icon-pen-to-square-solid";
-export * from "./icon-person-running-regular";
-export * from "./icon-play-solid";
-export * from "./icon-plug-regular";
-export * from "./icon-plus";
-export * from "./icon-rainbow-hash";
-export * from "./icon-rotate-regular";
-export * from "./icon-ruler-regular";
-export * from "./icon-shapes-regular";
-export * from "./icon-sidebar-regular";
-export * from "./icon-sparkles-light";
-export * from "./icon-star-regular";
-export * from "./icon-star-solid";
-export * from "./icon-table-light";
-export * from "./icon-terminal-light";
-export * from "./icon-thought-bubble-light";
-export * from "./icon-user-plus-regular";
-export * from "./icon-wand-magic-sparkles";
-export * from "./icon-x-twitter";
-export * from "./image-with-checked-background";
-export * from "./input-props";
-export * from "./loading-spinner";
-export * from "./menu-checkbox-item";
-export * from "./menu-item";
-export * from "./modal";
-export * from "./ontology-chip";
-export * from "./ontology-icons";
-export * from "./parse-url-for-ontology-chip";
-export * from "./pen-regular-icon";
-export * from "./popper-placement-modifier";
-export * from "./select";
-export * from "./selector-autocomplete";
-export * from "./skeleton";
-export * from "./submit-button-wrapper";
-export * from "./text-field";
-export * from "./type-card";
-export * from "./white-card";
+export * from "./alert-modal.js";
+export * from "./autocomplete.js";
+export * from "./autocomplete-dropdown.js";
+export * from "./avatar.js";
+export * from "./button.js";
+export * from "./callout.js";
+export * from "./chip.js";
+export * from "./chip-group.js";
+export * from "./data-type-selector.js";
+export * from "./e-chart.js";
+export * from "./entity-or-type-icon.js";
+export * from "./fa-icons/icons.js";
+export * from "./fontawesome-icon.js";
+export * from "./form-inline.js";
+export * from "./icon-angle-right-regular.js";
+export * from "./icon-arrow-down-left-and-arrow-up-right-to-center.js";
+export * from "./icon-arrow-down-regular.js";
+export * from "./icon-arrow-left.js";
+export * from "./icon-arrow-right-regular.js";
+export * from "./icon-arrow-rotate-left.js";
+export * from "./icon-arrow-up-regular.js";
+export * from "./icon-arrow-up-right.js";
+export * from "./icon-arrow-up-right-and-arrow-down-left-from-center.js";
+export * from "./icon-arrow-up-right-from-square-regular.js";
+export * from "./icon-arrow-up-right-regular.js";
+export * from "./icon-arrow-up-wide-short-light.js";
+export * from "./icon-arrows-rotate-regular.js";
+export * from "./icon-asterisk-regular.js";
+export * from "./icon-at-regular.js";
+export * from "./icon-barcode.js";
+export * from "./icon-bell-light.js";
+export * from "./icon-bug.js";
+export * from "./icon-bullseye-light.js";
+export * from "./icon-button.js";
+export * from "./icon-caret-down-solid.js";
+export * from "./icon-check.js";
+export * from "./icon-check-regular.js";
+export * from "./icon-chrome.js";
+export * from "./icon-circle-check-regular.js";
+export * from "./icon-circle-ellipsis-regular.js";
+export * from "./icon-circle-nodes-light.js";
+export * from "./icon-circle-one-regular.js";
+export * from "./icon-circle-plus.js";
+export * from "./icon-clock-regular.js";
+export * from "./icon-close.js";
+export * from "./icon-code.js";
+export * from "./icon-copy-regular.js";
+export * from "./icon-dash.js";
+export * from "./icon-diagram-nested-light.js";
+export * from "./icon-diagram-regular.js";
+export * from "./icon-discord.js";
+export * from "./icon-download-regular.js";
+export * from "./icon-eye-regular.js";
+export * from "./icon-eye-slash-regular.js";
+export * from "./icon-eye-solid.js";
+export * from "./icon-feather-regular.js";
+export * from "./icon-file-regular.js";
+export * from "./icon-file-spreadsheet-regular.js";
+export * from "./icon-file-spreadsheet-solid.js";
+export * from "./icon-forward-step-solid.js";
+export * from "./icon-graph.js";
+export * from "./icon-image-regular.js";
+export * from "./icon-image-solid.js";
+export * from "./icon-infinity-light.js";
+export * from "./icon-input-pipe.js";
+export * from "./icon-input-pipe-regular.js";
+export * from "./icon-lightbulb-on-rainbow.js";
+export * from "./icon-lightbulb-on-regular.js";
+export * from "./icon-link.js";
+export * from "./icon-list-regular.js";
+export * from "./icon-magnifying-glass-minus-light.js";
+export * from "./icon-magnifying-glass-plus-light.js";
+export * from "./icon-magnifying-glass-regular.js";
+export * from "./icon-memo-circle-check-regular.js";
+export * from "./icon-microscope-regular.js";
+export * from "./icon-pen-to-square-solid.js";
+export * from "./icon-person-running-regular.js";
+export * from "./icon-play-solid.js";
+export * from "./icon-plug-regular.js";
+export * from "./icon-plus.js";
+export * from "./icon-rainbow-hash.js";
+export * from "./icon-rotate-regular.js";
+export * from "./icon-ruler-regular.js";
+export * from "./icon-shapes-regular.js";
+export * from "./icon-sidebar-regular.js";
+export * from "./icon-sparkles-light.js";
+export * from "./icon-star-regular.js";
+export * from "./icon-star-solid.js";
+export * from "./icon-table-light.js";
+export * from "./icon-terminal-light.js";
+export * from "./icon-thought-bubble-light.js";
+export * from "./icon-user-plus-regular.js";
+export * from "./icon-wand-magic-sparkles.js";
+export * from "./icon-x-twitter.js";
+export * from "./image-with-checked-background.js";
+export * from "./input-props.js";
+export * from "./loading-spinner.js";
+export * from "./menu-checkbox-item.js";
+export * from "./menu-item.js";
+export * from "./modal.js";
+export * from "./ontology-chip.js";
+export * from "./ontology-icons.js";
+export * from "./parse-url-for-ontology-chip.js";
+export * from "./pen-regular-icon.js";
+export * from "./popper-placement-modifier.js";
+export * from "./select.js";
+export * from "./selector-autocomplete.js";
+export * from "./skeleton.js";
+export * from "./submit-button-wrapper.js";
+export * from "./text-field.js";
+export * from "./type-card.js";
+export * from "./white-card.js";
