+++ libs/@hashintel/design-system/src/data-type-selector.tsx
@@ -5,14 +5,7 @@ import type {
   StringConstraints,
   VersionedUrl,
 } from "@blockprotocol/type-system";
-import {
-  compareOntologyTypeVersions,
-  componentsFromVersionedUrl,
-} from "@blockprotocol/type-system";
-// eslint-disable-next-line no-restricted-imports -- TODO needs fixing if this package to be used from npm
-import type { ClosedDataTypeDefinition } from "@local/hash-graph-sdk/ontology";
-// eslint-disable-next-line no-restricted-imports -- TODO needs fixing if this package to be used from npm
-import { getMergedDataTypeSchema } from "@local/hash-isomorphic-utils/data-types";
+import { compareOntologyTypeVersions } from "@blockprotocol/type-system";
 import {
   Box,
   outlinedInputClasses,
@@ -25,18 +18,23 @@ import {
 import type { MouseEventHandler, ReactNode, RefObject } from "react";
 import { useEffect, useMemo, useRef, useState } from "react";
 
-import { getIconForDataType } from "./data-type-selector/icons";
-import { FontAwesomeIcon } from "./fontawesome-icon";
-import { IconButton } from "./icon-button";
-import { CaretDownSolidIcon } from "./icon-caret-down-solid";
-import { CheckIcon } from "./icon-check";
+import { getIconForDataType } from "./data-type-selector/icons.js";
+import { FontAwesomeIcon } from "./fontawesome-icon.js";
+import { IconButton } from "./icon-button.js";
+import { CaretDownSolidIcon } from "./icon-caret-down-solid.js";
+import { CheckIcon } from "./icon-check.js";
 
 export {
   getIconForDataType,
   identifierTypeTitles,
   measurementTypeTitles,
-} from "./data-type-selector/icons";
+} from "./data-type-selector/icons.js";
 
+/**
+ * Keep synced with the DataTypeForSelector type in @local/hash-isomorphic-utils/data-types.ts
+ *
+ * @todo create a 'shared frontend utils' package to avoid the need for duplication.
+ */
 export type DataTypeForSelector = {
   $id: VersionedUrl;
   abstract: boolean;
@@ -49,198 +47,6 @@ export type DataTypeForSelector = {
   type: string;
   title: string;
   version: OntologyTypeVersion;
-};
-
-const isDataType = (
-  dataType: DataType | ClosedDataTypeDefinition,
-): dataType is DataType => {
-  return "$id" in dataType;
-};
-
-const inputHeight = 48;
-const hintHeight = 36;
-
-/**
- * Build a tree rooted at the provided data type, with each node containing the information needed for the selector.
- *
- * Some callers will have the 'closed' data type (with all constraints resolved), others only the underlying schema,
- * without having its parents' constraints merged in.
- *
- * We don't care about the closed schema for this purpose because the selector doesn't care about value constraints,
- * but it's the structure that some callers may already have.
- */
-const transformDataTypeForSelector = <
-  T extends DataType | ClosedDataTypeDefinition,
->(
-  dataType: T,
-  directChildrenByDataTypeId: Record<VersionedUrl, T[]>,
-  allChildren: VersionedUrl[] = [],
-): {
-  allChildren: VersionedUrl[];
-  transformedDataType: DataTypeForSelector;
-} => {
-  const schema = isDataType(dataType) ? dataType : dataType.schema;
-  const { $id } = schema;
-
-  const children = directChildrenByDataTypeId[$id] ?? [];
-  const transformedChildren: DataTypeForSelector[] = [];
-
-  /**
-   * We need to track all descendants of the data type, so that {@link buildDataTypeTreesForSelector}
-   * can remove any types as roots of a tree where they appear lower down another tree.
-   */
-  allChildren.push(
-    ...children.map((child) =>
-      isDataType(child) ? child.$id : child.schema.$id,
-    ),
-  );
-
-  for (const child of children) {
-    const { transformedDataType } = transformDataTypeForSelector(
-      child,
-      directChildrenByDataTypeId,
-      allChildren,
-    );
-    transformedChildren.push(transformedDataType);
-  }
-
-  let format;
-  let type;
-
-  if (isDataType(dataType)) {
-    const firstSchema = "anyOf" in dataType ? dataType.anyOf[0] : dataType;
-    type = firstSchema.type;
-    format = "format" in firstSchema ? firstSchema.format : undefined;
-  } else {
-    const mergedSchema = getMergedDataTypeSchema(dataType.schema);
-
-    const firstSchema =
-      "anyOf" in mergedSchema ? mergedSchema.anyOf[0]! : mergedSchema;
-
-    type = firstSchema.type;
-    format = "format" in firstSchema ? firstSchema.format : undefined;
-  }
-
-  const { baseUrl, version } = componentsFromVersionedUrl($id);
-
-  const transformedDataType = {
-    $id,
-    baseUrl,
-    abstract: !!schema.abstract,
-    children: transformedChildren.sort((a, b) =>
-      a.title.localeCompare(b.title),
-    ),
-    description: schema.description,
-    directParents: isDataType(dataType)
-      ? (dataType.allOf?.map(({ $ref }) => $ref) ?? [])
-      : dataType.parents,
-    format,
-    label: schema.label,
-    title: schema.title,
-    type,
-    version,
-  };
-
-  return {
-    allChildren,
-    transformedDataType,
-  };
-};
-
-/**
- * Build trees of data types that can be selected by the user, rooted at those data types
- * among targetDataTypes which do not have any parents in targetDataTypes,
- * i.e. rooted at the types which have no selectable parents.
- *
- * Data types may appear in multiple locations in the tree if they have multiple selectable parents.
- * This will only happen if a user builds a type with multiple expected values,
- * and one of the values is a descendant of another one.
- *
- * The data type pool must include all children of the targetDataTypes, which can be fetched from the API via one of:
- * - fetching all data types:
- *     e.g. in the context of the type editor, where all types are selectable)
- * - making a query for an entityType with the 'resolvedWithDataTypeChildren' resolution method
- *     e.g. when selecting a value valid for specific entity types, and having the API resolve the valid data types
- */
-export const buildDataTypeTreesForSelector = <
-  T extends ClosedDataTypeDefinition | DataType,
->({
-  targetDataTypes,
-  dataTypePoolById,
-}: {
-  /**
-   * The data types that the user is allowed to select
-   * – does not need to include children, but they should appear in the dataTypePool.
-   */
-  targetDataTypes: T[];
-  /**
-   * All data types that are available for building the trees from.
-   * This MUST include targetDataTypes and any children of targetDataTypes to allow them to be included in the
-   * resulting tree. It MAY include other data types which are not selectable (e.g. parents of targetDataTypes, or
-   * unrelated types). Unrelated types will not be included in the resulting trees.
-   */
-  dataTypePoolById: Record<VersionedUrl, T>;
-}): DataTypeForSelector[] => {
-  const directChildrenByDataTypeId: Record<VersionedUrl, T[]> = {};
-
-  /**
-   * First, we need to know the children of all data types. Data types store references to their parents, not children.
-   * The selectable types are either targets or children of targets.
-   */
-  for (const dataType of Object.values(dataTypePoolById)) {
-    const directParents = isDataType(dataType)
-      ? (dataType.allOf?.map(({ $ref }) => $ref) ?? [])
-      : dataType.parents;
-
-    for (const parent of directParents) {
-      directChildrenByDataTypeId[parent] ??= [];
-
-      const parentDataType = dataTypePoolById[parent];
-
-      if (parentDataType) {
-        /**
-         * If the parentDataType is not in the pool, it is not a selectable parent.
-         * The caller is responsible for ensuring that the pool contains all selectable data types,
-         * via one of the methods described in the function's JSDoc.
-         */
-        directChildrenByDataTypeId[parent].push(dataType);
-      }
-    }
-  }
-
-  const rootsById: Record<VersionedUrl, DataTypeForSelector> = {};
-
-  const dataTypesBelowRoots: VersionedUrl[] = [];
-
-  /**
-   * Build a tree for each target data type.
-   */
-  for (const dataType of targetDataTypes) {
-    const { allChildren, transformedDataType } = transformDataTypeForSelector(
-      dataType,
-      directChildrenByDataTypeId,
-    );
-
-    if (
-      /**
-       * There's no point adding abstract types with no children, because they cannot be selected.
-       */
-      !transformedDataType.abstract ||
-      transformedDataType.children.length > 0
-    ) {
-      rootsById[transformedDataType.$id] = transformedDataType;
-      dataTypesBelowRoots.push(...allChildren);
-    }
-  }
-
-  /**
-   * Finally, remove any trees rooted at a target data type which appears as the child of another target.
-   */
-  for (const dataTypeId of dataTypesBelowRoots) {
-    delete rootsById[dataTypeId];
-  }
-
-  return Object.values(rootsById);
 };
 
 const DataTypeLabel = (props: {
@@ -597,6 +403,8 @@ export type DataTypeSelectorProps = {
 };
 
 const maxMenuHeight = 300;
+const inputHeight = 48;
+const hintHeight = 36;
 
 export const DataTypeSelector = (props: DataTypeSelectorProps) => {
   const {
