+++ libs/@hashintel/design-system/src/fa-icons/icons.tsx
@@ -1,22 +1,22 @@
-export { fa100 } from "./fa-100";
-export { faAsterisk } from "./fa-asterisk";
-export { faAtRegular } from "./fa-at-regular";
-export { faBracketsCurly } from "./fa-brackets-curly";
-export { faBracketsSquare } from "./fa-brackets-square";
-export { faCalendarClockRegular } from "./fa-calendar-clock-regular";
-export { faCalendarRegular } from "./fa-calendar-regular";
-export { faCheck } from "./fa-check";
-export { faCirclePlus } from "./fa-circle-plus";
-export { faClockRegular } from "./fa-clock-regular";
-export { faCube } from "./fa-cube";
-export { faCubes } from "./fa-cubes";
-export { faDiagramSubtask } from "./fa-diagram-subtask";
-export { faEmptySet } from "./fa-empty-set";
-export { faInputPipeRegular } from "./fa-input-pipe-regular";
-export { faListTree } from "./fa-list-tree";
-export { faPen } from "./fa-pen";
-export { faPlus } from "./fa-plus";
-export { faRulerRegular } from "./fa-ruler-regular";
-export { faSquareCheck } from "./fa-square-check";
-export { faText } from "./fa-text";
-export { faTrash } from "./fa-trash";
+export { fa100 } from "./fa-100.js";
+export { faAsterisk } from "./fa-asterisk.js";
+export { faAtRegular } from "./fa-at-regular.js";
+export { faBracketsCurly } from "./fa-brackets-curly.js";
+export { faBracketsSquare } from "./fa-brackets-square.js";
+export { faCalendarClockRegular } from "./fa-calendar-clock-regular.js";
+export { faCalendarRegular } from "./fa-calendar-regular.js";
+export { faCheck } from "./fa-check.js";
+export { faCirclePlus } from "./fa-circle-plus.js";
+export { faClockRegular } from "./fa-clock-regular.js";
+export { faCube } from "./fa-cube.js";
+export { faCubes } from "./fa-cubes.js";
+export { faDiagramSubtask } from "./fa-diagram-subtask.js";
+export { faEmptySet } from "./fa-empty-set.js";
+export { faInputPipeRegular } from "./fa-input-pipe-regular.js";
+export { faListTree } from "./fa-list-tree.js";
+export { faPen } from "./fa-pen.js";
+export { faPlus } from "./fa-plus.js";
+export { faRulerRegular } from "./fa-ruler-regular.js";
+export { faSquareCheck } from "./fa-square-check.js";
+export { faText } from "./fa-text.js";
+export { faTrash } from "./fa-trash.js";
