+++ libs/@hashintel/design-system/src/icon-circle-plus.tsx
@@ -1,5 +1,5 @@
-import { faCirclePlus } from "./fa-icons/icons";
-import { FontAwesomeIcon } from "./fontawesome-icon";
+import { faCirclePlus } from "./fa-icons/icons.js";
+import { FontAwesomeIcon } from "./fontawesome-icon.js";
 
 export const StyledPlusCircleIcon = () => (
   <FontAwesomeIcon
