+++ libs/@hashintel/design-system/src/text-field.tsx
@@ -10,8 +10,8 @@ import {
 import type { ReactNode } from "react";
 import { forwardRef, useState } from "react";
 
-import type { TextFieldProps } from "./input-props";
-import { getInputProps, inputLabelProps } from "./input-props";
+import type { TextFieldProps } from "./input-props.js";
+import { getInputProps, inputLabelProps } from "./input-props.js";
 
 /**
  * 'Freezes' a value when it's falsy, meaning the value will never update to
