+++ libs/@hashintel/design-system/src/theme/components.ts
@@ -6,10 +6,10 @@ import {
   MuiListItemSecondaryActionThemeOptions,
   MuiListItemTextThemeOptions,
   MuiTooltipThemeOptions,
-} from "./components/data-display";
-import { MuiChipThemeOptions } from "./components/data-display/mui-chip-theme-options";
-import { MuiListItemIconThemeOptions } from "./components/data-display/mui-list-item-icon-theme-options";
-import { MuiSkeletonThemeOptions } from "./components/feedback/mui-skeleton-theme-options";
+} from "./components/data-display.js";
+import { MuiChipThemeOptions } from "./components/data-display/mui-chip-theme-options.js";
+import { MuiListItemIconThemeOptions } from "./components/data-display/mui-list-item-icon-theme-options.js";
+import { MuiSkeletonThemeOptions } from "./components/feedback/mui-skeleton-theme-options.js";
 import {
   MuiButtonThemeOptions,
   MuiCheckboxThemeOptions,
@@ -18,18 +18,18 @@ import {
   MuiOutlinedInputThemeOptions,
   MuiRadioThemeOptions,
   MuiSwitchThemeOptions,
-} from "./components/inputs";
-import { MuiButtonBaseThemeOptions } from "./components/inputs/mui-button-base-theme-options";
-import { MuiInputBaseThemeOptions } from "./components/inputs/mui-input-base-theme-options";
-import { MuiSelectThemeOptions } from "./components/inputs/mui-select-theme-options";
+} from "./components/inputs.js";
+import { MuiButtonBaseThemeOptions } from "./components/inputs/mui-button-base-theme-options.js";
+import { MuiInputBaseThemeOptions } from "./components/inputs/mui-input-base-theme-options.js";
+import { MuiSelectThemeOptions } from "./components/inputs/mui-select-theme-options.js";
 import {
   MuiDrawerThemeOptions,
   MuiMenuItemThemeOptions,
   MuiMenuThemeOptions,
-} from "./components/navigation";
-import { MuiTabThemeOptions } from "./components/navigation/mui-tab-theme-options";
-import { MuiTabsThemeOptions } from "./components/navigation/mui-tabs-theme-options";
-import { MuiCssBaselineThemeOptions } from "./components/utils";
+} from "./components/navigation.js";
+import { MuiTabThemeOptions } from "./components/navigation/mui-tab-theme-options.js";
+import { MuiTabsThemeOptions } from "./components/navigation/mui-tabs-theme-options.js";
+import { MuiCssBaselineThemeOptions } from "./components/utils.js";
 
 export const components: Components<Theme> = {
   /** ===== INPUTS ===== */
