+++ libs/@hashintel/design-system/src/theme/components/data-display.ts
@@ -1,6 +1,6 @@
-export * from "./data-display/mui-icon-button-theme-options";
-export * from "./data-display/mui-list-item-button-theme-options";
-export * from "./data-display/mui-list-item-secondary-action-theme-options";
-export * from "./data-display/mui-list-item-text-theme-options";
-export * from "./data-display/mui-tooltip-theme-options";
-export * from "./data-display/mui-typography-theme-options";
+export * from "./data-display/mui-icon-button-theme-options.js";
+export * from "./data-display/mui-list-item-button-theme-options.js";
+export * from "./data-display/mui-list-item-secondary-action-theme-options.js";
+export * from "./data-display/mui-list-item-text-theme-options.js";
+export * from "./data-display/mui-tooltip-theme-options.js";
+export * from "./data-display/mui-typography-theme-options.js";
