+++ libs/@hashintel/design-system/src/theme/components/inputs.ts
@@ -1,7 +1,7 @@
-export * from "./inputs/mui-button-theme-options";
-export * from "./inputs/mui-checkbox-theme-options";
-export * from "./inputs/mui-form-helper-text-theme-options";
-export * from "./inputs/mui-input-label-theme-options";
-export * from "./inputs/mui-outlined-input-theme-options";
-export * from "./inputs/mui-radio-theme-options";
-export * from "./inputs/mui-switch-theme-options";
+export * from "./inputs/mui-button-theme-options.js";
+export * from "./inputs/mui-checkbox-theme-options.js";
+export * from "./inputs/mui-form-helper-text-theme-options.js";
+export * from "./inputs/mui-input-label-theme-options.js";
+export * from "./inputs/mui-outlined-input-theme-options.js";
+export * from "./inputs/mui-radio-theme-options.js";
+export * from "./inputs/mui-switch-theme-options.js";
