+++ libs/@hashintel/design-system/src/theme/components/inputs/mui-radio-theme-options.tsx
@@ -1,7 +1,7 @@
 import type { Components, Theme } from "@mui/material";
 
-import { RadioCheckedIcon } from "./mui-radio-theme-options/radio-checked-icon";
-import { RadioUncheckedIcon } from "./mui-radio-theme-options/radio-unchecked-icon";
+import { RadioCheckedIcon } from "./mui-radio-theme-options/radio-checked-icon.js";
+import { RadioUncheckedIcon } from "./mui-radio-theme-options/radio-unchecked-icon.js";
 
 const focusBorderOffset = 4;
 const focusBorderWidth = 2;
