+++ libs/@hashintel/design-system/src/theme/components/navigation.ts
@@ -1,3 +1,3 @@
-export * from "./navigation/mui-drawer-theme-options";
-export * from "./navigation/mui-menu-item-theme-options";
-export * from "./navigation/mui-menu-theme-options";
+export * from "./navigation/mui-drawer-theme-options.js";
+export * from "./navigation/mui-menu-item-theme-options.js";
+export * from "./navigation/mui-menu-theme-options.js";
